:bar_chart: Pass test to ensure the user can click on the plan my voyage btn
